Issues v0.0.1 Issues.MyTableFormatter

Functions

build_header(keys, column_sizes, pad_size)

build_rows(list, keys, column_sizes, pad_size)

build_table(header, rows)

format(list, keys, opts \\ [pad_size: 1])

Takes a list of rows as Maps and a list of strings, which references keys existents on Maps passed in the first list Returns a formatted table where the keys are setted as header columns followed by each row.

Example

iex> list = [
...>    %{
...>     "id" => "3",
...>      "created_at" => "2015-01-01T01:01:01Z",
...>      "title" => "foo bar baz",
...>      "other_data" => "other data"
...>    },
...>    %{
...>      "id" => "823",
...>      "created_at" => "2016-01-01T01:01:01Z",
...>      "title" => "foo bar baz quux",
...>      "other_data" => "other data"
...>    }
...>  ]
iex> Issues.MyTableFormatter.format(list, ["id", "created_at", "title"])
"""
| #   | created_at           | title            |
+-----+----------------------+------------------+
| 3   | 2015-01-01T01:01:01Z | foo bar baz      |
+-----+----------------------+------------------+
| 823 | 2016-01-01T01:01:01Z | foo bar baz quux |
+-----+----------------------+------------------+
"""